requirements.txt
flask
flask-cors
torch
torchvision
numpy
scikit-learn
matplotlib
seaborn
Pillow
joblib
python-dotenv
requests
tqdm
pandas
huggingface_hub
kaggle
